+++ backend/src/main/java/com/debede/backend/repository/Carrito_UsuarioRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Carrito_UsuarioRepository extends JpaRepository<Carrito_Usuario, Long> {
+public interface Carrito_UsuarioRepository extends JpaRepository<Carrito_Usuario, Integer> {
 
 }
